requirements.txt
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
pyscopg2==66.0.2
python-dotenv==1.0.1
SQLAlchemy==2.0.25
sqlparse==0.4.4
typing_extensions==4.9.0
Werkzeug==3.0.1
